inotify: Work around race condition by adding a retry loop
In certain circumstances usbmuxd might not have been started up when the
socket file creation event has occured. This causes connect_usbmuxd_socket()
to fail and usbmuxd_listen_inotify() is invoked again, but the socket file
creation event will not occur anymore. To fix this we retry to connect to
usbmuxd after waiting a second in case the first connection attempt failed
(with a maximum of 10 retries).