Visual C++ compatibility: fix C2375
Visual C++ requires the declaration of methods to be exactly the same in the .h and .c files; otherwise it throws C2375.

This is fixed by:
- Defining USBMUXD_API_MSC in the public headers (via usbmuxd.h) which will add __declspec( dllexport ) to all method definitions in the public headers when compiling with Visual C++
- Ensuring USBMUXD_API_MSC is always set to __declspec( dllexport ) in the private source files when compiling with Visual C++.

Using USBMUXD_API in the public headers does not work, becuase its value depends on HAVE_FVISIBILITY, which may be defined in config.h, which is not a public header.