socket: Fix socket_connect_addr() not connecting using IPv6 in some cases
This extends the socket helper with functions to determine the "scope" and a
suitable "scope id" of an IPv6 address. While socket_connect_addr() prefers
any initially supplied "scope id" to maintain routing information if possible,
it will attempt to determine the best suitable route with the new helpers.
This became a requirement during testing with remote usbmux connections that
provide a different "scope id" and thus might cause IPv6 routing to not work
at all. Thus the "scope id" is only valid per host.